chore( working through some default values in images and card. Also breaking out Galleries from Images )
