+++ package.json
@@ -37,13 +37,13 @@
   },
   "devDependencies": {
     "@storybook/addon-actions": "^5.2.1",
-    "@storybook/addon-docs": "^5.3.0-alpha.2",
+    "@storybook/addon-docs": "^5.3.0-alpha.3",
     "@storybook/addon-knobs": "^5.2.1",
     "@storybook/addon-links": "^5.2.1",
     "@storybook/react": "^5.2.1",
     "@storybook/storybook-deployer": "^2.8.1",
     "@storybook/addons": "^5.2.0-beta.32",
-    "@storybook/addon-storysource": "^5.3.0-alpha.2",
+    "@storybook/addon-storysource": "^5.3.0-alpha.3",
     "@storybook/source-loader": "^5.2.0-beta.32",
     "babel-cli": "^6.26.0",
     "babel-eslint": "^10.0.3",
